Split component loader in 2
